Added CxPybind target for common code between plugins
* Added initial docs generation
