+++ conda/meta.yaml
@@ -45,6 +45,7 @@ requirements:
   host:
     - python
     - numpy >=1.16
+    - mypy
     - boost_mp11
     - eigen
     - fmt =10
