+++ conda/recipe.yaml
@@ -47,6 +47,7 @@ requirements:
     - git
   host:
     - python
+    - mypy
     - numpy >=1.16
     - boost_mp11
     - eigen
