+++ wrapping/python/docs/api.rst
@@ -1,0 +1,9 @@
+complex
+===
+
+.. autosummary::
+   :toctree: generated
+   :recursive:
+   :imported-members:
+
+   complex
