+++ wrapping/python/docs/index.rst
@@ -1,0 +1,9 @@
+complex Documentation
+============================
+
+Contents:
+
+.. toctree::
+   :maxdepth: 2
+
+   api
